requirements.txt
Flask
flask_cors
google-genai